% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Activities}
\alias{Activities}
\title{Activities Object}
\usage{
Activities(filters = NULL, metricNames = NULL)
}
\arguments{
\item{filters}{List of activity filters}

\item{metricNames}{List of names of floodlight activity metrics}
}
\value{
Activities object
}
\description{
Activities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an activity group.
}

