% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AdvertisersListResponse}
\alias{AdvertisersListResponse}
\title{AdvertisersListResponse Object}
\usage{
AdvertisersListResponse(advertisers = NULL, nextPageToken = NULL)
}
\arguments{
\item{advertisers}{Advertiser collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
AdvertisersListResponse object
}
\description{
AdvertisersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Advertiser List Response
}

