% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{campaignCreativeAssociations.list}
\alias{campaignCreativeAssociations.list}
\title{Retrieves the list of creative IDs associated with the specified campaign.}
\usage{
campaignCreativeAssociations.list(profileId, campaignId, maxResults = NULL,
  pageToken = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{campaignId}{Campaign ID in this association}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

