% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ChangeLog}
\alias{ChangeLog}
\title{ChangeLog Object}
\usage{
ChangeLog(accountId = NULL, action = NULL, changeTime = NULL,
  fieldName = NULL, id = NULL, newValue = NULL, objectId = NULL,
  objectType = NULL, oldValue = NULL, subaccountId = NULL,
  transactionId = NULL, userProfileId = NULL, userProfileName = NULL)
}
\arguments{
\item{accountId}{Account ID of the modified object}

\item{action}{Action which caused the change}

\item{changeTime}{Time when the object was modified}

\item{fieldName}{Field name of the object which changed}

\item{id}{ID of this change log}

\item{newValue}{New value of the object field}

\item{objectId}{ID of the object of this change log}

\item{objectType}{Object type of the change log}

\item{oldValue}{Old value of the object field}

\item{subaccountId}{Subaccount ID of the modified object}

\item{transactionId}{Transaction ID of this change log}

\item{userProfileId}{ID of the user who modified the object}

\item{userProfileName}{User profile name of the user who modified the object}
}
\value{
ChangeLog object
}
\description{
ChangeLog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a change that a user has made to a resource.
}

