% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ChangeLogsListResponse}
\alias{ChangeLogsListResponse}
\title{ChangeLogsListResponse Object}
\usage{
ChangeLogsListResponse(changeLogs = NULL, nextPageToken = NULL)
}
\arguments{
\item{changeLogs}{Change log collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
ChangeLogsListResponse object
}
\description{
ChangeLogsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Change Log List Response
}

