% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ClickTag}
\alias{ClickTag}
\title{ClickTag Object}
\usage{
ClickTag(eventName = NULL, name = NULL, value = NULL)
}
\arguments{
\item{eventName}{Advertiser event name associated with the click tag}

\item{name}{Parameter name for the specified click tag}

\item{value}{Parameter value for the specified click tag}
}
\value{
ClickTag object
}
\description{
ClickTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Click Tag.
}

