% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Creative}
\alias{Creative}
\title{Creative Object}
\usage{
Creative(accountId = NULL, active = NULL, adParameters = NULL,
  adTagKeys = NULL, advertiserId = NULL, allowScriptAccess = NULL,
  archived = NULL, artworkType = NULL, authoringSource = NULL,
  authoringTool = NULL, auto_advance_images = NULL,
  backgroundColor = NULL, backupImageClickThroughUrl = NULL,
  backupImageFeatures = NULL, backupImageReportingLabel = NULL,
  backupImageTargetWindow = NULL, clickTags = NULL, commercialId = NULL,
  companionCreatives = NULL, compatibility = NULL,
  convertFlashToHtml5 = NULL, counterCustomEvents = NULL,
  creativeAssets = NULL, creativeFieldAssignments = NULL,
  customKeyValues = NULL, exitCustomEvents = NULL, fsCommand = NULL,
  htmlCode = NULL, htmlCodeLocked = NULL, id = NULL,
  idDimensionValue = NULL, lastModifiedInfo = NULL,
  latestTraffickedCreativeId = NULL, name = NULL, overrideCss = NULL,
  redirectUrl = NULL, renderingId = NULL,
  renderingIdDimensionValue = NULL, requiredFlashPluginVersion = NULL,
  requiredFlashVersion = NULL, size = NULL, skippable = NULL,
  sslCompliant = NULL, sslOverride = NULL, studioAdvertiserId = NULL,
  studioCreativeId = NULL, studioTraffickedCreativeId = NULL,
  subaccountId = NULL, thirdPartyBackupImageImpressionsUrl = NULL,
  thirdPartyRichMediaImpressionsUrl = NULL, thirdPartyUrls = NULL,
  timerCustomEvents = NULL, totalFileSize = NULL, type = NULL,
  version = NULL, videoDescription = NULL, videoDuration = NULL)
}
\arguments{
\item{accountId}{Account ID of this creative}

\item{active}{Whether the creative is active}

\item{adParameters}{Ad parameters user for VPAID creative}

\item{adTagKeys}{Keywords for a Rich Media creative}

\item{advertiserId}{Advertiser ID of this creative}

\item{allowScriptAccess}{Whether script access is allowed for this creative}

\item{archived}{Whether the creative is archived}

\item{artworkType}{Type of artwork used for the creative}

\item{authoringSource}{Source application where creative was authored}

\item{authoringTool}{Authoring tool for HTML5 banner creatives}

\item{auto_advance_images}{Whether images are automatically advanced for enhanced image creatives}

\item{backgroundColor}{The 6-character HTML color code, beginning with #, for the background of the window area where the Flash file is displayed}

\item{backupImageClickThroughUrl}{Click-through URL for backup image}

\item{backupImageFeatures}{List of feature dependencies that will cause a backup image to be served if the browser that serves the ad does not support them}

\item{backupImageReportingLabel}{Reporting label used for HTML5 banner backup image}

\item{backupImageTargetWindow}{Target window for backup image}

\item{clickTags}{Click tags of the creative}

\item{commercialId}{Industry standard ID assigned to creative for reach and frequency}

\item{companionCreatives}{List of companion creatives assigned to an in-Stream videocreative}

\item{compatibility}{Compatibilities associated with this creative}

\item{convertFlashToHtml5}{Whether Flash assets associated with the creative need to be automatically converted to HTML5}

\item{counterCustomEvents}{List of counter events configured for the creative}

\item{creativeAssets}{Assets associated with a creative}

\item{creativeFieldAssignments}{Creative field assignments for this creative}

\item{customKeyValues}{Custom key-values for a Rich Media creative}

\item{exitCustomEvents}{List of exit events configured for the creative}

\item{fsCommand}{OpenWindow FSCommand of this creative}

\item{htmlCode}{HTML code for the creative}

\item{htmlCodeLocked}{Whether HTML code is DCM-generated or manually entered}

\item{id}{ID of this creative}

\item{idDimensionValue}{Dimension value for the ID of this creative}

\item{lastModifiedInfo}{Creative last modification information}

\item{latestTraffickedCreativeId}{Latest Studio trafficked creative ID associated with rich media and VPAID creatives}

\item{name}{Name of the creative}

\item{overrideCss}{Override CSS value for rich media creatives}

\item{redirectUrl}{URL of hosted image or hosted video or another ad tag}

\item{renderingId}{ID of current rendering version}

\item{renderingIdDimensionValue}{Dimension value for the rendering ID of this creative}

\item{requiredFlashPluginVersion}{The minimum required Flash plugin version for this creative}

\item{requiredFlashVersion}{The internal Flash version for this creative as calculated by DoubleClick Studio}

\item{size}{Size associated with this creative}

\item{skippable}{Whether the user can choose to skip the creative}

\item{sslCompliant}{Whether the creative is SSL-compliant}

\item{sslOverride}{Whether creative should be treated as SSL compliant even if the system scan shows it's not}

\item{studioAdvertiserId}{Studio advertiser ID associated with rich media and VPAID creatives}

\item{studioCreativeId}{Studio creative ID associated with rich media and VPAID creatives}

\item{studioTraffickedCreativeId}{Studio trafficked creative ID associated with rich media and VPAID creatives}

\item{subaccountId}{Subaccount ID of this creative}

\item{thirdPartyBackupImageImpressionsUrl}{Third-party URL used to record backup image impressions}

\item{thirdPartyRichMediaImpressionsUrl}{Third-party URL used to record rich media impressions}

\item{thirdPartyUrls}{Third-party URLs for tracking in-stream video creative events}

\item{timerCustomEvents}{List of timer events configured for the creative}

\item{totalFileSize}{Combined size of all creative assets}

\item{type}{Type of this creative}

\item{version}{The version number helps you keep track of multiple versions of your creative in your reports}

\item{videoDescription}{Description of the video ad}

\item{videoDuration}{Creative video duration in seconds}
}
\value{
Creative object
}
\description{
Creative Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a Creative.
}
\seealso{
Other Creative functions: \code{\link{creatives.insert}},
  \code{\link{creatives.patch}},
  \code{\link{creatives.update}}
}

