% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{creativeFieldValues.update}
\alias{creativeFieldValues.update}
\title{Updates an existing creative field value.}
\usage{
creativeFieldValues.update(CreativeFieldValue, profileId, creativeFieldId)
}
\arguments{
\item{CreativeFieldValue}{The \link{CreativeFieldValue} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{creativeFieldId}{Creative field ID for this creative field value}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other CreativeFieldValue functions: \code{\link{CreativeFieldValue}},
  \code{\link{creativeFieldValues.insert}},
  \code{\link{creativeFieldValues.patch}}
}

