% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DimensionValueList}
\alias{DimensionValueList}
\title{DimensionValueList Object}
\usage{
DimensionValueList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{The eTag of this response for caching purposes}

\item{items}{The dimension values returned in this response}

\item{nextPageToken}{Continuation token used to page through dimension values}
}
\value{
DimensionValueList object
}
\description{
DimensionValueList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the list of DimensionValue resources.
}

