% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{EventTag}
\alias{EventTag}
\title{EventTag Object}
\usage{
EventTag(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, campaignId = NULL,
  campaignIdDimensionValue = NULL, enabledByDefault = NULL,
  excludeFromAdxRequests = NULL, id = NULL, name = NULL,
  siteFilterType = NULL, siteIds = NULL, sslCompliant = NULL,
  status = NULL, subaccountId = NULL, type = NULL, url = NULL,
  urlEscapeLevels = NULL)
}
\arguments{
\item{accountId}{Account ID of this event tag}

\item{advertiserId}{Advertiser ID of this event tag}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{campaignId}{Campaign ID of this event tag}

\item{campaignIdDimensionValue}{Dimension value for the ID of the campaign}

\item{enabledByDefault}{Whether this event tag should be automatically enabled for all of the advertiser's campaigns and ads}

\item{excludeFromAdxRequests}{Whether to remove this event tag from ads that are trafficked through DoubleClick Bid Manager to Ad Exchange}

\item{id}{ID of this event tag}

\item{name}{Name of this event tag}

\item{siteFilterType}{Site filter type for this event tag}

\item{siteIds}{Filter list of site IDs associated with this event tag}

\item{sslCompliant}{Whether this tag is SSL-compliant or not}

\item{status}{Status of this event tag}

\item{subaccountId}{Subaccount ID of this event tag}

\item{type}{Event tag type}

\item{url}{Payload URL for this event tag}

\item{urlEscapeLevels}{Number of times the landing page URL should be URL-escaped before being appended to the click-through event tag URL}
}
\value{
EventTag object
}
\description{
EventTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of an event tag.
}
\seealso{
Other EventTag functions: \code{\link{eventTags.insert}},
  \code{\link{eventTags.patch}},
  \code{\link{eventTags.update}}
}

