% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivityGroupsListResponse}
\alias{FloodlightActivityGroupsListResponse}
\title{FloodlightActivityGroupsListResponse Object}
\usage{
FloodlightActivityGroupsListResponse(floodlightActivityGroups = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{floodlightActivityGroups}{Floodlight activity group collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
FloodlightActivityGroupsListResponse object
}
\description{
FloodlightActivityGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Floodlight Activity Group List Response
}

