% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{inventoryItems.list}
\alias{inventoryItems.list}
\title{Retrieves a list of inventory items, possibly filtered.}
\usage{
inventoryItems.list(profileId, projectId, ids = NULL, inPlan = NULL,
  maxResults = NULL, orderId = NULL, pageToken = NULL, siteId = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{projectId}{Project ID for order documents}

\item{ids}{Select only inventory items with these IDs}

\item{inPlan}{Select only inventory items that are in plan}

\item{maxResults}{Maximum number of results to return}

\item{orderId}{Select only inventory items that belong to specified orders}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{siteId}{Select only inventory items that are associated with these sites}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

