% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{InventoryItemsListResponse}
\alias{InventoryItemsListResponse}
\title{InventoryItemsListResponse Object}
\usage{
InventoryItemsListResponse(inventoryItems = NULL, nextPageToken = NULL)
}
\arguments{
\item{inventoryItems}{Inventory item collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
InventoryItemsListResponse object
}
\description{
InventoryItemsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inventory item List Response
}

