% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OperatingSystemsListResponse}
\alias{OperatingSystemsListResponse}
\title{OperatingSystemsListResponse Object}
\usage{
OperatingSystemsListResponse(operatingSystems = NULL)
}
\arguments{
\item{operatingSystems}{Operating system collection}
}
\value{
OperatingSystemsListResponse object
}
\description{
OperatingSystemsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Operating System List Response
}

