% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OptimizationActivity}
\alias{OptimizationActivity}
\title{OptimizationActivity Object}
\usage{
OptimizationActivity(floodlightActivityId = NULL,
  floodlightActivityIdDimensionValue = NULL, weight = NULL)
}
\arguments{
\item{floodlightActivityId}{Floodlight activity ID of this optimization activity}

\item{floodlightActivityIdDimensionValue}{Dimension value for the ID of the floodlight activity}

\item{weight}{Weight associated with this optimization}
}
\value{
OptimizationActivity object
}
\description{
OptimizationActivity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative optimization activity.
}

