% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementsGenerateTagsResponse}
\alias{PlacementsGenerateTagsResponse}
\title{PlacementsGenerateTagsResponse Object}
\usage{
PlacementsGenerateTagsResponse(placementTags = NULL)
}
\arguments{
\item{placementTags}{Set of generated tags for the specified placements}
}
\value{
PlacementsGenerateTagsResponse object
}
\description{
PlacementsGenerateTagsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement GenerateTags Response
}

