% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Recipient}
\alias{Recipient}
\title{Recipient Object}
\usage{
Recipient(deliveryType = NULL, email = NULL)
}
\arguments{
\item{deliveryType}{The delivery type for the recipient}

\item{email}{The email address of the recipient}
}
\value{
Recipient object
}
\description{
Recipient Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a recipient.
}

