% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(Report.criteria = NULL, Report.crossDimensionReachCriteria = NULL,
  Report.delivery = NULL, Report.floodlightCriteria = NULL,
  Report.floodlightCriteria.reportProperties = NULL,
  Report.pathToConversionCriteria = NULL,
  Report.pathToConversionCriteria.reportProperties = NULL,
  Report.reachCriteria = NULL, Report.schedule = NULL, accountId = NULL,
  criteria = NULL, crossDimensionReachCriteria = NULL, delivery = NULL,
  etag = NULL, fileName = NULL, floodlightCriteria = NULL,
  format = NULL, id = NULL, lastModifiedTime = NULL, name = NULL,
  ownerProfileId = NULL, pathToConversionCriteria = NULL,
  reachCriteria = NULL, schedule = NULL, subAccountId = NULL,
  type = NULL)
}
\arguments{
\item{Report.criteria}{The \link{Report.criteria} object or list of objects}

\item{Report.crossDimensionReachCriteria}{The \link{Report.crossDimensionReachCriteria} object or list of objects}

\item{Report.delivery}{The \link{Report.delivery} object or list of objects}

\item{Report.floodlightCriteria}{The \link{Report.floodlightCriteria} object or list of objects}

\item{Report.floodlightCriteria.reportProperties}{The \link{Report.floodlightCriteria.reportProperties} object or list of objects}

\item{Report.pathToConversionCriteria}{The \link{Report.pathToConversionCriteria} object or list of objects}

\item{Report.pathToConversionCriteria.reportProperties}{The \link{Report.pathToConversionCriteria.reportProperties} object or list of objects}

\item{Report.reachCriteria}{The \link{Report.reachCriteria} object or list of objects}

\item{Report.schedule}{The \link{Report.schedule} object or list of objects}

\item{accountId}{The account ID to which this report belongs}

\item{criteria}{The report criteria for a report of type 'STANDARD'}

\item{crossDimensionReachCriteria}{The report criteria for a report of type 'CROSS_DIMENSION_REACH'}

\item{delivery}{The report's email delivery settings}

\item{etag}{The eTag of this response for caching purposes}

\item{fileName}{The filename used when generating report files for this report}

\item{floodlightCriteria}{The report criteria for a report of type 'FLOODLIGHT'}

\item{format}{The output format of the report}

\item{id}{The unique ID identifying this report resource}

\item{lastModifiedTime}{The timestamp (in milliseconds since epoch) of when this report was last modified}

\item{name}{The name of the report}

\item{ownerProfileId}{The user profile id of the owner of this report}

\item{pathToConversionCriteria}{The report criteria for a report of type 'PATH_TO_CONVERSION'}

\item{reachCriteria}{The report criteria for a report of type 'REACH'}

\item{schedule}{The report's schedule}

\item{subAccountId}{The subaccount ID to which this report belongs if applicable}

\item{type}{The type of the report}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Report resource.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}

