% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RichMediaExitOverride}
\alias{RichMediaExitOverride}
\title{RichMediaExitOverride Object}
\usage{
RichMediaExitOverride(customExitUrl = NULL, exitId = NULL,
  useCustomExitUrl = NULL)
}
\arguments{
\item{customExitUrl}{Click-through URL to override the default exit URL}

\item{exitId}{ID for the override to refer to a specific exit in the creative}

\item{useCustomExitUrl}{Whether to use the custom exit URL}
}
\value{
RichMediaExitOverride object
}
\description{
RichMediaExitOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rich Media Exit Override.
}

