% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRolePermission}
\alias{UserRolePermission}
\title{UserRolePermission Object}
\usage{
UserRolePermission(availability = NULL, id = NULL, name = NULL,
  permissionGroupId = NULL)
}
\arguments{
\item{availability}{Levels of availability for a user role permission}

\item{id}{ID of this user role permission}

\item{name}{Name of this user role permission}

\item{permissionGroupId}{ID of the permission group that this user role permission belongs to}
}
\value{
UserRolePermission object
}
\description{
UserRolePermission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a user role permission.
}

